[KSP-70] 게이트웨이 MS에서 프로젝트 영문 명 수정
com.psear -> com.pser
