--- src/main/java/com/psear/gateway/common/response/ApiResponse.java
@@ -1,19 +1,0 @@
-package com.psear.gateway.common.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ApiResponse<T> {
-    private String message = "";
-    private T body = null;
-
-    public static ApiResponse<Void> error(String message) {
-        return new ApiResponse<>(message, null);
-    }
-
-    public static <T> ApiResponse<T> success(T body) {
-        return new ApiResponse<>("", body);
-    }
-}
