--- src/main/java/com/psear/gateway/config/ExceptionConfig.java
@@ -1,19 +1,0 @@
-package com.psear.gateway.config;
-
-import com.psear.gateway.exception.GlobalExceptionHandler;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.web.reactive.error.ErrorWebExceptionHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class ExceptionConfig {
-    private final ObjectMapper objectMapper;
-
-    @Bean
-    public ErrorWebExceptionHandler globalExceptionHandler() {
-        return new GlobalExceptionHandler(objectMapper);
-    }
-}
