--- src/main/java/com/psear/gateway/domain/RoleWritingConverter.java
@@ -1,13 +1,0 @@
-package com.psear.gateway.domain;
-
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.data.convert.WritingConverter;
-import org.springframework.lang.NonNull;
-
-@WritingConverter
-public class RoleWritingConverter implements Converter<RoleEnum, Integer> {
-    @Override
-    public Integer convert(@NonNull RoleEnum source) {
-        return source.getValue();
-    }
-}
