--- src/main/java/com/psear/gateway/filter/TrailingSlashRedirectFilter.java
@@ -1,25 +1,0 @@
-package com.psear.gateway.filter;
-
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.lang.NonNull;
-import org.springframework.web.server.ServerWebExchange;
-import org.springframework.web.server.WebFilter;
-import org.springframework.web.server.WebFilterChain;
-import reactor.core.publisher.Mono;
-
-public class TrailingSlashRedirectFilter implements WebFilter {
-    @Override
-    @NonNull
-    public Mono<Void> filter(ServerWebExchange exchange, @NonNull WebFilterChain chain) {
-        ServerHttpRequest request = exchange.getRequest();
-        String path = request.getPath().value();
-
-        if (path.endsWith("/") && !path.equals("/")) {
-            String newPath = path.substring(0, path.length() - 1);
-            ServerHttpRequest newRequest = request.mutate().path(newPath).build();
-            return chain.filter(exchange.mutate().request(newRequest).build());
-        }
-
-        return chain.filter(exchange);
-    }
-}
