+++ src/main/java/com/pser/gateway/GatewayApplication.java
@@ -1,0 +1,14 @@
+package com.pser.gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.reactive.config.EnableWebFlux;
+
+@SpringBootApplication
+@EnableWebFlux
+public class GatewayApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GatewayApplication.class, args);
+    }
+}
