+++ src/main/java/com/pser/gateway/config/CorsConfig.java
@@ -1,0 +1,27 @@
+package com.pser.gateway.config;
+
+import java.util.List;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.reactive.CorsConfigurationSource;
+import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
+
+@Configuration
+public class CorsConfig {
+
+    @Bean("corsConfigurationSource")
+    CorsConfigurationSource corsConfigurationSource() {
+        List<String> allowedOrigins = List.of("http://localhost:3000");
+        List<String> allowedMethods = List.of("*");
+        List<String> allowedHeaders = List.of("*");
+
+        CorsConfiguration configuration = new CorsConfiguration();
+        configuration.setAllowedOrigins(allowedOrigins);
+        configuration.setAllowedMethods(allowedMethods);
+        configuration.setAllowedHeaders(allowedHeaders);
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration);
+        return source;
+    }
+}
