+++ src/main/java/com/pser/gateway/config/ExceptionConfig.java
@@ -1,0 +1,19 @@
+package com.pser.gateway.config;
+
+import com.pser.gateway.exception.GlobalExceptionHandler;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.web.reactive.error.ErrorWebExceptionHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+public class ExceptionConfig {
+    private final ObjectMapper objectMapper;
+
+    @Bean
+    public ErrorWebExceptionHandler globalExceptionHandler() {
+        return new GlobalExceptionHandler(objectMapper);
+    }
+}
