+++ src/main/java/com/pser/gateway/domain/BaseEntity.java
@@ -1,0 +1,43 @@
+package com.pser.gateway.domain;
+
+import java.time.LocalDateTime;
+import java.util.Objects;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Column;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class BaseEntity {
+    @Id
+    private Long id;
+
+    @Column("createdAt")
+    private LocalDateTime createdAt;
+
+    @Column("updatedAt")
+    private LocalDateTime updatedAt;
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (!this.getClass().equals(obj.getClass())) {
+            return false;
+        }
+        BaseEntity baseEntity = (BaseEntity) obj;
+        return Objects.equals(this.id, baseEntity.id);
+    }
+
+    @Override
+    public int hashCode() {
+        if (id == null) {
+            return 0;
+        }
+        return id.hashCode();
+    }
+}
