+++ src/main/java/com/pser/gateway/domain/RoleWritingConverter.java
@@ -1,0 +1,13 @@
+package com.pser.gateway.domain;
+
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.data.convert.WritingConverter;
+import org.springframework.lang.NonNull;
+
+@WritingConverter
+public class RoleWritingConverter implements Converter<RoleEnum, Integer> {
+    @Override
+    public Integer convert(@NonNull RoleEnum source) {
+        return source.getValue();
+    }
+}
