+++ src/main/java/com/pser/gateway/dto/UserDetailsImpl.java
@@ -1,0 +1,18 @@
+package com.pser.gateway.dto;
+
+import java.util.Collection;
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+
+@Getter
+public class UserDetailsImpl extends User {
+    private final Long id;
+
+    public UserDetailsImpl(Long id, String username, String password,
+                           Collection<? extends GrantedAuthority> authorities) {
+        super(username, password, authorities);
+        this.id = id;
+    }
+}
