--- src/test/java/com/psear/gateway/GatewayApplicationTests.java
@@ -1,13 +1,0 @@
-package com.psear.gateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GatewayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
