+++ src/test/java/com/pser/gateway/GatewayApplicationTests.java
@@ -1,0 +1,13 @@
+package com.pser.gateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GatewayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
